[detail city] animation from shared element image works
