+++ ComfortZoneApp/app/src/main/java/com/example/comfortzone/GetWeatherCallback.java
