+++ ComfortZoneApp/app/src/main/res/layout/activity_city_detail.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CityDetailFragment"
+    tools:context=".CityDetailActivity"
     android:orientation="vertical">
