+++ ComfortZoneApp/app/src/main/res/values/strings.xml
@@ -28,4 +28,5 @@
     <string name="comfort">Comfort:</string>
     <string name="temperature">Temperature:</string>
     <string name="cityIcon">cityIcon</string>
+    <string name="cityId">cityId</string>
 </resources>
